backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
requests
openai-whisper
python-multipart
